+++ src/main/kotlin/cloud/glitchdev/rfu/feature/debug/Entities.kt
@@ -1,0 +1,13 @@
+package cloud.glitchdev.rfu.feature.debug
+
+import cloud.glitchdev.rfu.utils.command.AbstractCommand
+
+object Entities : AbstractCommand("entities") {
+    override val description: String = "Commands for entity debugging (sb, bobbers, normal)"
+
+    init {
+        append(Sb)
+        append(Bobbers)
+        append(Normal)
+    }
+}