+++ CHANGELOG.md
@@ -9,3 +9,6 @@
 the resource pack is not active to prevent the missing textures with it not loaded
 - Fixed infinite loading screen when downloading the server resource pack for the first time
 - Fixed resource pack triggering a reload even though it wasn't needed
+
+### Back-end
+- Added internal tracking to link sea creatures to their originating fishing bobber